// src/timesheet-web/src/app/api/models/task-item-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface TaskItemDto {
  assigneeUserId?: number | null;
  backlogId?: number;
  id?: number;
  isActive?: boolean;
  orderIndex?: number;
  rowVersion?: number;
  status?: string | null;
  taskName?: string | null;
  type?: string | null;
}